erase

+++ src/data.cpp
@@ -39,4 +39,4 @@ int yTilesOffset	= yTilesFont + fontSize + 10;
 //Arrays
 std::vector<Texture2D> entities;
 std::vector<Texture2D> tiles;
-std::vector<int> canvas(canvasTileWidth * canvasTileWidth, 200);
+std::vector<u16> canvas(canvasTileWidth * canvasTileWidth, 200);
